tweet_scraper/requirements.txt
python-dotenv
requests
pandas
pyarrow